+++ test/#2.test.ts
@@ -1,39 +1,36 @@
-import { StructBuffer, sview, WORD } from "../src";
+import { StructBuffer, sview, uint16_t } from "../src";
 
 // https://github.com/januwA/struct-buffer/issues/2
 
 describe("debug", () => {
   it("string_t", () => {
-    const s = new StructBuffer(
-      "test",
-      {
-        a: WORD,
-        b: WORD,
-        c: new StructBuffer(
-          "test2",
-          {
-            ip: WORD,
-            port: WORD,
-          },
-          {
-            littleEndian: false,
-          }
-        ),
-      },
-      {
-        littleEndian: true,
-      }
-    );
+    const s = new StructBuffer({
+      a: uint16_t,
+      b: uint16_t,
+      c: new StructBuffer({
+        ip: uint16_t,
+        port: uint16_t,
+      }),
+    });
 
-    const v = s.encode({
+    const obj = {
       a: 1,
       b: 2,
       c: {
         ip: 10,
         port: 100,
       },
+    };
+
+    const v1 = s.encode(obj, {
+      littleEndian: false,
     });
-    expect(sview(v)).toBe("01 00 02 00 00 0a 00 64");
+    expect(sview(v1)).toBe("00 01 00 02 00 0a 00 64");
+
+    const v2 = s.encode(obj, {
+      littleEndian: true,
+    });
+    expect(sview(v2)).toBe("01 00 02 00 0a 00 64 00");
   });
 
   it("test extends", () => {
@@ -41,7 +38,7 @@ describe("debug", () => {
       a = 10;
     }
 
-    let s = new X("s", {})[2][2] as any;
+    let s = new X({})[2][2] as any;
     expect(s.a).toBe(10);
   });
 });